% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersections.R
\name{find_intersections}
\alias{find_intersections}
\title{Intersect a Line with Spatial Polygons and Compute Segment Lengths}
\usage{
find_intersections(polys, line_proj)
}
\arguments{
\item{polys}{An \code{sf} object representing spatial polygons, such as intensity zones or regions.}

\item{line_proj}{An \code{sf} object representing a projected line geometry, typically the result of transforming a great-circle line between two locations to match the CRS of \code{polys}.}
}
\value{
An \code{sf} object representing the intersected segments, with an added column \code{seg_length} indicating the length of each segment in kilometers.
}
\description{
Identifies and retains portions of a line that intersect with spatial polygons,
computes the lengths of the intersected segments, and filters out zero-length segments.
}
\keyword{internal}
