% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descrp.R
\docType{data}
\name{pgv}
\alias{pgv}
\alias{pga}
\alias{psa0p1}
\alias{psa0p3}
\alias{psa1}
\alias{psa3}
\title{Ground Motion Polygon Datasets}
\format{
Each dataset is an \code{sf} object with polygons and associated parameters.

An object of class \code{sf} (inherits from \code{data.frame}) with 322 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 322 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 322 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 322 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 322 rows and 9 columns.
}
\source{
Derived from regional simulations.
}
\usage{
pgv

pga

psa0p1

psa0p3

psa1

psa3
}
\description{
Datasets containing spatial correlation model outputs for PGV, PGA, and PSA values.
}
\keyword{datasets}
